+++ src/ScottPlot/Enums/ArrowAnchor.cs
@@ -1,0 +1,20 @@
+﻿namespace ScottPlot
+{
+    public enum ArrowAnchor
+    {
+        /// <summary>
+        /// X/Y coordinates define the base of the arrow
+        /// </summary>
+        Base,
+
+        /// <summary>
+        /// X/Y coordinates define the center of the arrow
+        /// </summary>
+        Center,
+
+        /// <summary>
+        /// X/Y coordinates define the tip of the arrow
+        /// </summary>
+        Tip,
+    }
+}