+++ src/demo/ScottPlot.Demo.WPF/WpfDemos/MultiAxisLock.xaml.cs
@@ -70,6 +70,7 @@ namespace ScottPlot.Demo.WPF.WpfDemos
 
             // adjust axis limits to fit the data once before locking them
             WpfPlot1.Plot.AxisAuto();
+            WpfPlot1.Render();
             CheckChanged(null, null);
         }