+++ src/demo/ScottPlot.Demo.WPF/WpfDemos/ShowValueOnHover.xaml.cs
@@ -39,6 +39,9 @@ namespace ScottPlot.Demo.WPF.WpfDemos
             HighlightedPoint.MarkerSize = 10;
             HighlightedPoint.MarkerShape = ScottPlot.MarkerShape.openCircle;
             HighlightedPoint.IsVisible = false;
+
+            // perform an initial render
+            wpfPlot1.Render();
         }
 
         private void wpfPlot1_MouseMove(object sender, MouseEventArgs e)