+++ dev/changelog.md
@@ -4,6 +4,13 @@
 * Ticks: Improve placement when axis scale lock is enabled (#1229, #1197)
 * Plot: `SetViewLimits()` replaced by `SetOuterViewLimits()` and `SetInnerViewLimits()` (#1197) _Thanks @noob765_
 * Plot: `EqualScaleMode` (an enumeration accepted by `AxisScaleLock()`) now has `PreserveSmallest` and `PreserveLargest` members to indicate which axis to prioritize when adjusting zoom level. The new default is `PreserveSmallest` which prevents data from falling off the edge of the plot when resizing. (#1197) _Thanks @noob765_
+* Axis: Improved alignment of 90º rotated ticks (#1194, #1201) _Thanks @gigios_
+* Controls: Fix bug where middle-click-drag zoom rectangle would persist if combined with scroll wheel events (#1226) _Thanks @Elgot_
+* Scatter Plot: Fixed bug affecting plots where `YError` is set but `XError` is not (#1237, #1238) _Thanks @simmdan_
+* Palette: Added `Microcharts` colorset (#1235) _Thanks @arthurits_
+* SignalPlotXY: Added support for `FillType` (#1232) _Thanks @ddrrrr_
+* Arrow: New plot type for rendering arrows on plots. Arrowhead functionality of scatter plots has been deprecated. (#1241, #1240)
+* Controls: Automatic rendering has been deprecated. Users must call Render() manually at least once. (#1165, #1117)
 * Radial Gauge Plots: `AddRadialGauge()` now adds a radial gauge plot (a new circular plot type where values are represented as arcs spanning a curve). See cookbook for examples and documentation. (#1242) _Thanks @arthurits_
 
 ## ScottPlot 4.1.17